% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGLMM.R
\name{calc_r_tau2}
\alias{calc_r_tau2}
\title{Godambe calibration}
\usage{
calc_r_tau2(
  var,
  init,
  tau2h,
  mll,
  tau2.min = 1e-06,
  diff_step = 0.001,
  cap = 50,
  smooth = 1
)
}
\description{
Godambe calibration
}
