% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{chu2020}
\alias{chu2020}
\title{dataset of Chu et al. (2020)}
\format{
A \code{data.frame} for binary distributed outcome with 7 studies:
\describe{
\item{Study}{Study ID}
\item{x1}{Events of Covid-19 for group of further distnce}
\item{n1}{Number of subjects of Covid-19 for group of further distnce}
\item{x0}{Events of Covid-19 for group of shorter distnce}
\item{n0}{Number of subjects of Covid-19 for group of shorter distnce}
}
}
\source{
Chu, D. K., Akl, E. A., Duda, S., Solo, K., Yaacoub, S., Schünemann, H. J., ... & Reinap, M. (2020). Physical distancing, face masks, and eye protection to prevent person-to-person transmission of SARS-CoV-2 and COVID-19: a systematic review and meta-analysis. \emph{The lancet}, 395(10242), 1973-1987.
Data used under the terms of the Creative Commons Attribution IGO 3.0 License (CC BY 3.0 IGO):
\url{https://creativecommons.org/licenses/by/3.0/igo/}
}
\usage{
chu2020
}
\description{
dataset of Chu et al. (2020)
}
\keyword{datasets}
