% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGLMM.R
\name{confint_AN}
\alias{confint_AN}
\title{confidence interval for metaGLMM object using normal approximation.}
\usage{
confint_AN(
  object,
  parm = names(coef(object))[-length(coef(object))],
  level = 0.95
)
}
\arguments{
\item{object}{object of metaGLMM function.}

\item{parm}{a vector of parameter names for estimating confidence interval.}

\item{level}{a value of significance level. The default is 0.95.}
}
\value{
a matrix of confidence interval using normal approximation.
}
\description{
confidence interval for metaGLMM object using normal approximation.
}
