% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGLMM.R
\name{confint_PL}
\alias{confint_PL}
\title{confidence interval for metaGLMM object using profile likelihood method.}
\usage{
confint_PL(
  object,
  parm = names(coef(object))[-length(coef(object))],
  level = 0.95,
  renge.c = 30
)
}
\arguments{
\item{object}{object of metaGLMM function.}

\item{parm}{a vector of parameter names for estimating confidence interval.}

\item{level}{a value of significance level. The default is 0.95.}

\item{renge.c}{a value of search renge. The default is 30.}
}
\value{
a matrix of confidence interval using profile likelihood method.
}
\description{
confidence interval for metaGLMM object using profile likelihood method.
}
