% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{likelihood function}
\usage{
likelihood(
  formula,
  data,
  vi,
  ni,
  beta,
  tau2,
  family = binomial(link = "logit"),
  tau2.min = 1e-06,
  re_group = NULL,
  trt = NULL,
  rstdnorm = qnorm((qrng::sobol(5000, d = 1, scrambling = 1) * (5000 - 1) + 0.5)/5000)
)
}
\arguments{
\item{formula}{formula of meta-analysis.}

\item{data}{data of meta-analysis.}

\item{vi}{variances in each study.}

\item{ni}{subjects in each study.}

\item{beta}{a vector of parameter.}

\item{tau2}{a value of between-study variance.}

\item{family}{a family objects for models.}

\item{tau2.min}{minimum value of tau2 > 0.}

\item{rstdnorm}{random numbers. The default is 5000 quasi-random values by Sobol' quasi-random sequences}
}
\value{
value of likelihood function
}
\description{
likelihood function
}
