% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{logSumExp_simple}
\alias{logSumExp_simple}
\title{simple logSumExp funciton}
\usage{
logSumExp_simple(x)
}
\arguments{
\item{x}{value of between-study variance.}
}
\value{
value of logSumExp function.
}
\description{
simple logSumExp funciton
}
