% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{long2020}
\alias{long2020}
\title{dataset of Long et al. (2020)}
\format{
A \code{data.frame} for gamma distributed outcome with 5 studies:
\describe{
\item{Study}{Study ID}
\item{mi}{Mean of intensive care unit (ICU) length of stay (LoS)}
\item{sdi}{Standard deviation of ICU LoS}
\item{ni}{Number of subjects}
\item{trt}{1 is Treatment, an 0 is control}
}
}
\source{
Long, R., Tian, J., Wu, S., Li, Y., Yang, X., & Fei, J. (2020). Clinical efficacy of surgical versus conservative treatment for multiple rib fractures: a meta-analysis of randomized controlled trials. \emph{International Journal of Surgery}, 83, 79-88.
Data used under the terms of the Creative Commons Attribution 4.0 International License (CC BY 4.0):
\url{https://creativecommons.org/licenses/by/4.0/}
}
\usage{
long2020
}
\description{
dataset of Long et al. (2020)
}
\keyword{datasets}
