% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{make_ll_fun}
\alias{make_ll_fun}
\title{likelihood function}
\usage{
make_ll_fun(
  formula,
  data,
  vi,
  ni,
  tau2,
  family,
  tau2_var = FALSE,
  rstdnorm = qnorm((qrng::sobol(5000, d = 1, scrambling = 1) * (5000 - 1) + 0.5)/5000),
  ...
)
}
\arguments{
\item{formula}{formula of meta-analysis.}

\item{data}{data of meta-analysis.}

\item{vi}{variances in each study.}

\item{ni}{subjects in each study.}

\item{tau2}{a value of between-study variance.}

\item{family}{a family objects for models.}

\item{tau2_var}{If FALSE (default), tau2 is estimated. If TRUE, assume tau2 is true value.}

\item{rstdnorm}{random numbers. The default is 5000 quasi-random values by Sobol' quasi-random sequences}
}
\value{
function of likelihood
}
\description{
likelihood function
}
