% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGLMM-package.R
\docType{package}
\name{metaGLMM-package}
\alias{metaGLMM-package}
\alias{metaGLMM}
\title{metaGLMM}
\description{
meta-analysis using generalized linear mixed effects model with aggregate data
}
\references{
none
}
\author{
\strong{Maintainer}: Keisuke Hanada \email{keisuke.hanada@gmail.com}

}
