% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGLMM.R
\name{profile_ll}
\alias{profile_ll}
\title{profile likelihood function for metaGLMM object.}
\usage{
profile_ll(value, object, parm)
}
\arguments{
\item{value}{a vector of parameter values.}

\item{object}{object of metaGLMM function.}

\item{parm}{parameter name for profile likelihood function.}
}
\value{
a values of profile likelihood function.
}
\description{
profile likelihood function for metaGLMM object.
}
