% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGLMM.R
\name{profile_ll_sbc}
\alias{profile_ll_sbc}
\title{profile likelihood with simple Bartlett correction for metaGLMM object.}
\usage{
profile_ll_sbc(value, object, parm)
}
\arguments{
\item{value}{a vector of parameter values.}

\item{object}{object of metaGLMM function.}

\item{parm}{parameter name for profile likelihood function.}
}
\value{
a values of profile likelihood function.
}
\description{
profile likelihood with simple Bartlett correction for metaGLMM object.
}
