% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{rutter2021}
\alias{rutter2021}
\title{dataset of Rutter et al. (2021)}
\format{
A \code{data.frame} for Poisson distributed outcome with 11 studies:
\describe{
\item{Study}{Study ID}
\item{Year}{Publishing year}
\item{Sex}{Sex (Female or Male)}
\item{Events}{Number of events}
\item{Time}{Total observation time}
}
}
\source{
Rutter, M., Bowley, J., Lanyon, P. C., Grainge, M. J., & Pearce, F. A. (2021). A systematic review and meta-analysis of the incidence rate of Takayasu arteritis. \emph{Rheumatology}, 60(11), 4982-4990.
Data used under the terms of the Creative Commons Attribution 4.0 International License (CC BY 4.0):
\url{https://creativecommons.org/licenses/by/4.0/}
}
\usage{
rutter2021
}
\description{
dataset of Rutter et al. (2021)
}
\keyword{datasets}
