% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGLMM.R
\name{metaGLMM}
\alias{metaGLMM}
\title{meta-analysis using generalized linear mixed effects model with aggregate data}
\usage{
metaGLMM(
  formula,
  data,
  vi,
  ni,
  tau2,
  family,
  tau2_var = TRUE,
  rstdnorm = qnorm((qrng::sobol(5000, d = 1, scrambling = 1) * (5000 - 1) + 0.5)/5000)
)
}
\arguments{
\item{formula}{formula of meta-analysis.}

\item{data}{data of meta-analysis.}

\item{vi}{variances in each study.}

\item{ni}{subjects in each study.}

\item{tau2}{a value of between-study variance.}

\item{family}{a family objects for models.}

\item{tau2_var}{If TRUE (default), tau2 is estimated. If FALSE, assume tau2 is true value.}

\item{rstdnorm}{random numbers. The default is 5000 quasi-random values by Sobol' quasi-random sequences}
}
\value{
object of meta-analysis under generalized linear mixed effects model
}
\description{
meta-analysis using generalized linear mixed effects model with aggregate data
}
